Modified tasks.py so that user can do CRUD on their tasks only.Deleted in TaskUpdate class (schemas.py) id && user_id

+++ app/routes/tasks.py
@@ -27,17 +27,20 @@ user_dependency = Annotated[str, Depends(get_current_user)]
 
 
 @router.get('/task/{task_id}')
-async def get_task(user : user_dependency, db: db_dependency, task_id : int = Path(gt=0)):
+async def get_task(user : user_dependency, db: db_dependency, task_id : int):
     
     if user is None:
          raise HTTPException(status_code=401, detail='Authentication failed')
     
-    task_to_return = db.query(Task).filter(Task.id == task_id, Task.owner_id == user.get('id')).first()
+    task_to_return = db.query(Task).filter(Task.id == task_id).first()
 
     if not task_to_return:
         raise HTTPException(status_code=404, detail="Task not found")
     
-    return task_to_return
+    if task_to_return.owner_id == user.get('id') or user.get('role') == 'admin':
+        return task_to_return
+    else:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     
 
 @router.post('/task', status_code=status.HTTP_201_CREATED)
@@ -51,6 +54,7 @@ async def create_task(user : user_dependency, db : db_dependency, task_create : 
         raise HTTPException(status_code=400, detail="Task already exists")
 
     task_model = Task(**task_create.model_dump(), owner_id=user.get('id'))
+
 
     db.add(task_model)
     db.commit()
@@ -66,6 +70,9 @@ async def update_task(user : user_dependency, db : db_dependency, task_update : 
 
     if task_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found" )
+    
+    if task_model.owner_id != user.get('id') and user.get('role') != 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     
     update_data = task_update.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -86,6 +93,9 @@ async def delete_task(user : user_dependency, db : db_dependency, task_id : int)
 
     if task_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Task not found")
+    
+    if task_model.owner_id != user.get('id') and user.get('role') != 'admin':
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not enough permissions")
     
     db.delete(task_model)
     db.commit()
